build & install

+++ base/palloc.c
@@ -49,7 +49,7 @@ void *palloc(mem_pool *pool, size_t size)
 
         last = align_ptr(p->last, MEM_POOL_ALIGNMENT);
 
-        if (p->end - last >= size) {
+        if (p->end - last >= (long) size) {
             p->last = last + size;
             return last;
         }
